requirements_cuda.txt
--extra-index-url https://download.pytorch.org/whl/cu128

bcrypt>=5.0.0
python-dotenv>=0.9.9
flask>=3.1.3
joblib>=1.5.3
ollama>=0.6.1
pandas>=3.0.2
protobuf>=3.19.0
psycopg2>=2.9.11
requests>=2.33.1
tensorflow>=2.21.0
ultralytics>=8.4.33
torch>=2.7.0+cu128
torchvision>=0.22.0+cu128
torchaudio>=2.7.0+cu128
sentence-transformers>=5.3.0
chromadb>=1.5.5
opentelemetry-api>=1.40.0
opentelemetry-sdk>=1.40.0
opentelemetry-exporter-otlp-proto-grpc>=1.40.0
openpyxl>=3.1.5
git+https://github.com/openai/CLIP.git
ftfy>=6.3.1
regex>=2026.4.4
tqdm>=4.67.3